requirements/base.txt
Flask==1.0.2

phonenumbers==8.10.6
environs==4.1.0
mysqlclient==1.4.2
python-dateutil==2.8.0
csvvalidator==1.2
uWSGI==2.0.18

Flask-CORS==3.0.7
flask-expects-json==1.3.1
Flask-JWT-Extended==3.17.0
Flask-Bcrypt==0.7.1
Flask-Minio==0.1.2
flasgger==0.9.2

SQLAlchemy==1.2.18
SQLAlchemy-Utils==0.33.11
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.4.0

flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.16.0
